+++ examples/example.rs
@@ -3,6 +3,8 @@ extern crate shell2batch;
 fn main() {
     let script = shell2batch::convert(
         r#"
+        set -x
+
         export FILE1=file1
         export FILE2=file2
 
@@ -30,6 +32,8 @@ fn main() {
     assert_eq!(
         script,
         r#"
+@echo on
+
 set FILE1=file1
 set FILE2=file2