+++ examples/example.rs
@@ -3,6 +3,8 @@ extern crate shell2batch;
 fn main() {
     let script = shell2batch::convert(
         r#"
+        set -x
+
         export FILE1=file1
         export FILE2=file2
 
@@ -20,6 +22,8 @@ fn main() {
 
         unset MY_DIR
 
+        touch ./file3
+
         #provide custom windows command for specific shell command
         complex_bash_command --flag1 value2 # shell2batch: complex_windows_command /flag10 windows_value
         "#,
@@ -28,6 +32,8 @@ fn main() {
     assert_eq!(
         script,
         r#"
+@echo on
+
 set FILE1=file1
 set FILE2=file2
 
@@ -44,6 +50,8 @@ set MY_DIR=directory
 rmdir /S /Q %MY_DIR%
 
 set MY_DIR=
+
+copy /B .\file3+,, .\file3
 
 @REM provide custom windows command for specific shell command
 complex_windows_command /flag10 windows_value