+++ src/lib.rs
@@ -22,7 +22,6 @@
     legacy_constructor_visibility,
     legacy_directory_ownership,
     macro_expanded_macro_exports_accessed_by_absolute_paths,
-    meta_variable_misuse,
     missing_copy_implementations,
     missing_docs,
     missing_fragment_specifier,
@@ -183,8 +182,6 @@
 #[path = "./lib_test.rs"]
 mod lib_test;
 
-extern crate regex;
-
 mod converter;
 
 /// Converts the provided shell script and returns the windows batch script text.
@@ -197,6 +194,8 @@ mod converter;
 /// fn main() {
 ///     let script = shell2batch::convert(
 ///         r#"
+///         set -x
+///
 ///         export FILE1=file1
 ///         export FILE2=file2
 ///
@@ -214,6 +213,8 @@ mod converter;
 ///
 ///         unset MY_DIR
 ///
+///         touch ./file3
+///
 ///         #provide custom windows command for specific shell command
 ///         complex_bash_command --flag1 value2 # shell2batch: complex_windows_command /flag10 windows_value
 ///         "#
@@ -222,6 +223,8 @@ mod converter;
 ///     assert_eq!(
 ///         script,
 ///         r#"
+/// @echo on
+///
 /// set FILE1=file1
 /// set FILE2=file2
 ///
@@ -238,6 +241,8 @@ mod converter;
 /// rmdir /S /Q %MY_DIR%
 ///
 /// set MY_DIR=
+///
+/// copy /B .\file3+,, .\file3
 ///
 /// @REM provide custom windows command for specific shell command
 /// complex_windows_command /flag10 windows_value