+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### v0.4.2
 
-* Adding touch support #9
+* Add support for touch command #9
+* Add support for echo on/off #10
 
 ### v0.4.1 (2019-07-23)